+++ source/LearningHub.Model/Models/Course/CourseAgeDetail.cs
@@ -4,6 +4,6 @@
     {
         public int MinimumAge { get; set; }
         public int MaximumAge { get; set; }
-        public int AverageAge { get; set; }
+        public double AverageAge { get; set; }
     }
 }
